% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_freq_blue.R
\name{a_freq_blue}
\alias{a_freq_blue}
\title{from a long-format data frame of genotypes and a pedigree, estimate allele freq via BLUE}
\usage{
a_freq_blue(D, P)
}
\arguments{
\item{D}{data frame of genotypes.  Must have the columns ID, locus, gene_copy, and allele.
The names of the individuals in ID must match the names of the same individuals
in the pedgiree P.}

\item{P}{a data frame describing a pedigree with columns id, mom, dad, and sex, like one
returned by \code{\link{read_colony_best_config}}.}
}
\description{
Simple implementation of...
}

