% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported_utilities.R
\name{add_founder_rows}
\alias{add_founder_rows}
\title{add founders to pedigree as needed}
\usage{
add_founder_rows(D)
}
\arguments{
\item{D}{a data frame with columns id, mom, dad, and sex}
}
\description{
add founders to pedigree as needed
}

