% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_L_from_pedigree.R
\name{matrix_L_from_pedigree}
\alias{matrix_L_from_pedigree}
\title{given a pedigree as a data frame return the L matrix}
\usage{
matrix_L_from_pedigree(D, Ids)
}
\arguments{
\item{D}{a data frame with columns id, mom, dad, and sex that can be read
into kinship2's pedigree function}

\item{Ids}{the IDs/labels of the individuals you want the matrix to refer to
(i.e. the sampled individuals.)}
}
\description{
given a pedigree as a data frame return the L matrix
}
\examples{
example(read_colony_best_config)
# here are the ones we are interested in
Ids <- bcped$id[!is.na(bcped$dad)]
Lmat <- matrix_L_from_pedigree(bcped, Ids)
}

