% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_z.R
\name{optimal_z}
\alias{optimal_z}
\title{given a kinship-derived correlation matrix L, find the optimal sibship elimination}
\usage{
optimal_z(L)
}
\arguments{
\item{L}{a matrix like that returned by \code{\link{matrix_L_from_pedigree}}.}
}
\description{
This uses a greedy algorithm that is not necessarily optimal, but is likely to be
pretty close.
}

