% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_colony_best_config.R
\name{read_colony_best_config}
\alias{read_colony_best_config}
\title{read a Colony BestConfig file into a pedigree with columns id, mom, dad, sex}
\usage{
read_colony_best_config(path)
}
\arguments{
\item{path}{the path to the BestConfig file you want to read.}
}
\description{
read a Colony BestConfig file into a pedigree with columns id, mom, dad, sex
}
\examples{
bcfile <- system.file(file.path("extdata", "output.BestConfig"), package = "afblue")
bcped <- read_colony_best_config(bcfile)
}

