% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slg_pipe2long.R
\name{slg_pipe2long}
\alias{slg_pipe2long}
\title{read an slg_pipe format file and turn it into a long format data frame of genotypes}
\usage{
slg_pipe2long(S)
}
\arguments{
\item{S}{the path to the slg_pipe format file}
}
\description{
more later
}
\examples{
# this is not portable, just here for me at the moment
sfile <- "/Users/eriq/Documents/git-repos/ccc-sonc-resample/slg_pipe/arena/COHO_FIRST_RUN/genos_slg_pipe.txt"
slg_pipe2long(sfile)
}

