% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_sib_yank.R
\name{traditional_sib_yank}
\alias{traditional_sib_yank}
\title{given a one generation pedigree returns individuals you will keep after a "traditional" sib-yank}
\usage{
traditional_sib_yank(ped, Asize = 2, NumLeft = 1)
}
\arguments{
\item{ped}{A pedigree with colums at least of  id, mom, dad.  If mom or dad is NA, those rows are
removed.  (so you can use the result of \code{\link{read_colony_best_config}}).}

\item{Asize}{yank individuals out of full sibships that are strictly larger than this}

\item{NumLeft}{in a sibship of size larger than Asize, NumLeft indivdiuals will be left}
}
\value{
Returns a vector of id labels that will be kept
}
\description{
This does what we did in the NorCal steelhead paper.  For any inferred sibship of size strictly greater
then Asize, remove all the individuals except for NumLeft of them randomly chosen.  The defaults
reflect what we have done in the past.
}

