% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_matrix_L.R
\name{weights_from_matrix_L}
\alias{weights_from_matrix_L}
\title{Given a matrix L, determine the weights to be given to each individual's genotype}
\usage{
weights_from_matrix_L(Lmat)
}
\arguments{
\item{L}{a matrix, such as one that comes out of matrix_L_from_pedigree.  It should
have row and column names and be invertible}
}
\description{
Given a matrix L, determine the weights to be given to each individual's genotype
}
\examples{
example(matrix_L_from_pedigree)
wts <- weights_from_matrix_L(Lmat)
}

